GameLogic is added as a Dll plugin
The integration works as follows. The project has to be compiled into a
dll and put in the Assets/Plugins folder. With this Unity can use it as
a plugin. Visual studio is set to compile it into that folder so it is
_somewhat_ integrated.

It is not ideal, but Unity is somewhat crazy that way