Added initial Entity management
Entities are unaware of where they are (though hexagons aren't).
I made a static check so that Entities can only be added through the
map. This is needed because the map has a dictionary with the mapping
from (X,Y) -> Entity (could very well be Hexagon -> Entity, but this is
more general. Ensuring that creating and modification is through the
map helps ensure that the table is up to date with the actual data.