Added BFS Area functionality
Now you can with a map and an entity it's current movement list.
The code is abstracted so that if we want an arbitrary BFS, we can
use the same functionality underneath

TODO: The same error reporting mechanism seems to be all over the place,
perhaps is better to unify it in a standard place and all the useful
information there.