Added initial Action execution handling
This enables the view (unity) to tell the game when a certain action
was actually executing, so that it can actually do the effects of the
action. For now the only action is moving.

There is basic checking that the action list has to be completely
executed before other actions are to be run.